requirements.txt
aiogram==3.11.0
python-decouple~=3.8
telebot~=0.0.5
requests~=2.31.0
langchain~=0.3.9